Move types from rrweb-snapshot to @rrweb/types
