+++ packages/rrdom-nodejs/tsconfig.json
@@ -10,7 +10,7 @@
       "path": "../rrdom"
     },
     {
-      "path": "../rrweb-snapshot"
+      "path": "../types"
     }
   ]
 }
