+++ packages/rrweb-snapshot/package.json
@@ -52,6 +52,9 @@
     "url": "https://github.com/rrweb-io/rrweb/issues"
   },
   "homepage": "https://github.com/rrweb-io/rrweb/tree/master/packages/rrweb-snapshot#readme",
+  "dependencies": {
+    "@rrweb/types": "^2.0.0-alpha.14"
+  },
   "devDependencies": {
     "@types/jsdom": "^20.0.0",
     "@types/node": "^18.15.11",
