+++ packages/rrweb-snapshot/src/types.ts
@@ -1,77 +1,4 @@
-export enum NodeType {
-  Document,
-  DocumentType,
-  Element,
-  Text,
-  CDATA,
-  Comment,
-}
-
-export type documentNode = {
-  type: NodeType.Document;
-  childNodes: serializedNodeWithId[];
-  compatMode?: string;
-};
-
-export type documentTypeNode = {
-  type: NodeType.DocumentType;
-  name: string;
-  publicId: string;
-  systemId: string;
-};
-
-export type attributes = {
-  [key: string]: string | number | true | null;
-};
-export type legacyAttributes = {
-  /**
-   * @deprecated old bug in rrweb was causing these to always be set
-   * @see https://github.com/rrweb-io/rrweb/pull/651
-   */
-  selected: false;
-};
-
-export type elementNode = {
-  type: NodeType.Element;
-  tagName: string;
-  attributes: attributes;
-  childNodes: serializedNodeWithId[];
-  isSVG?: true;
-  needBlock?: boolean;
-  // This is a custom element or not.
-  isCustom?: true;
-};
-
-export type textNode = {
-  type: NodeType.Text;
-  textContent: string;
-  isStyle?: true;
-};
-
-export type cdataNode = {
-  type: NodeType.CDATA;
-  textContent: '';
-};
-
-export type commentNode = {
-  type: NodeType.Comment;
-  textContent: string;
-};
-
-export type serializedNode = (
-  | documentNode
-  | documentTypeNode
-  | elementNode
-  | textNode
-  | cdataNode
-  | commentNode
-) & {
-  rootId?: number;
-  isShadowHost?: boolean;
-  isShadow?: boolean;
-};
-
-export type serializedNodeWithId = serializedNode & { id: number };
+import type { serializedNodeWithId, NodeType } from '@rrweb/types';
 
 export type serializedElementNodeWithId = Extract<
   serializedNodeWithId,
@@ -82,56 +9,15 @@ export type tagMap = {
   [key: string]: string;
 };
 
-export type mediaAttributes = {
-  rr_mediaState: 'played' | 'paused';
-  rr_mediaCurrentTime: number;
-  /**
-   * for backwards compatibility this is optional but should always be set
-   */
-  rr_mediaPlaybackRate?: number;
-  /**
-   * for backwards compatibility this is optional but should always be set
-   */
-  rr_mediaMuted?: boolean;
-  /**
-   * for backwards compatibility this is optional but should always be set
-   */
-  rr_mediaLoop?: boolean;
-  /**
-   * for backwards compatibility this is optional but should always be set
-   */
-  rr_mediaVolume?: number;
-};
-
-// @deprecated
+/**
+ * @deprecated
+ */
 export interface INode extends Node {
   __sn: serializedNodeWithId;
 }
 
 export interface ICanvas extends HTMLCanvasElement {
   __context: string;
-}
-
-export interface IMirror<TNode> {
-  getId(n: TNode | undefined | null): number;
-
-  getNode(id: number): TNode | null;
-
-  getIds(): number[];
-
-  getMeta(n: TNode): serializedNodeWithId | null;
-
-  removeNodeFromMap(n: TNode): void;
-
-  has(id: number): boolean;
-
-  hasNode(node: TNode): boolean;
-
-  add(n: TNode, meta: serializedNodeWithId): void;
-
-  replace(id: number, n: TNode): void;
-
-  reset(): void;
 }
 
 export type idNodeMap = Map<number, Node>;
