+++ packages/rrweb-snapshot/test/rebuild.test.ts
@@ -9,7 +9,7 @@ import {
   buildNodeWithSN,
   createCache,
 } from '../src/rebuild';
-import { NodeType } from '../src/types';
+import { NodeType } from '@rrweb/types';
 import { createMirror, Mirror } from '../src/utils';
 
 function getDuration(hrtime: [number, number]) {
