+++ packages/rrweb-snapshot/test/utils.test.ts
@@ -2,9 +2,9 @@
  * @vitest-environment jsdom
  */
 import { describe, it, test, expect } from 'vitest';
-import { NodeType, serializedNode } from '../src/types';
+import { NodeType, serializedNode } from '@rrweb/types';
 import { extractFileExtension, isNodeMetaEqual } from '../src/utils';
-import type { serializedNodeWithId } from 'rrweb-snapshot';
+import type { serializedNodeWithId } from '@rrweb/types';
 
 describe('utils', () => {
   describe('isNodeMetaEqual()', () => {
