+++ packages/rrweb/src/record/workers/image-bitmap-data-url-worker.ts
@@ -1,6 +1,6 @@
 import { encode } from 'base64-arraybuffer';
-import type { DataURLOptions } from 'rrweb-snapshot';
 import type {
+  DataURLOptions,
   ImageBitmapDataURLWorkerParams,
   ImageBitmapDataURLWorkerResponse,
 } from '@rrweb/types';
