+++ packages/rrweb/src/replay/media/index.ts
@@ -1,8 +1,8 @@
 import { type Emitter, MediaInteractions, ReplayerEvents } from '@rrweb/types';
 import type { RRMediaElement } from 'rrdom';
 import type { createPlayerService, createSpeedService } from '../machine';
-import type { Mirror, mediaAttributes } from 'rrweb-snapshot';
-import type { mediaInteractionData } from '@rrweb/types';
+import type { Mirror } from 'rrweb-snapshot';
+import type { mediaInteractionData, mediaAttributes } from '@rrweb/types';
 
 type MediaState = {
   isPlaying: boolean;
